# swagger.yaml
openapi: 3.0.0
info:
  title: PSGC API
  version: 0.0.1
  description: API for listing all the regions, provinces, cities, municipalities, and barangays of the Philippines. <br/><br/>Current data is as of September 30, 2024 and is based from https://psa.gov.ph/classification/psgc. <br/><br/> For issues, features, and contributions, please visit the [GitHub repository](https://github.com/jeffreybernadas/psgc-api).
  contact:
    name: Jeffrey Bernadas
    email: 'jeffrey.bernadas0@gmail.com'
...
servers:
  - url: https://psgc.thecodebit.digital/
    description: 'PSGC API'
  - url: http://localhost:4000
    description: 'PSGC API: Local'
...
paths:
...
  /api/v1/city/{cityCode}/barangays:
    get:
      tags:
        - City
      summary: Get all barangays in a city.
      parameters:
        - in: path
          name: cityCode
          schema:
            type: string
            example: 1380100000
          description: The 10-digit PSGC code of the city.
        - in: query
          name: page
          schema:
            type: number
            default: 1
          description: The page number of the response.
        - in: query
          name: limit
          schema:
            type: number
            default: 10
          description: The number of items per page.
      responses:
        200:
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseSchemaBarangay'
...
  /api/v1/municipality/{municipalityCode}/barangays:
    get:
      tags:
        - Municipality
      summary: Get all barangays in a municipality.
      parameters:
        - in: path
          name: municipalityCode
          schema:
            type: string
            example: 1381701000
          description: The 10-digit PSGC code of the municipality.
        - in: query
          name: page
          schema:
            type: number
            default: 1
          description: The page number of the response.
        - in: query
          name: limit
          schema:
            type: number
            default: 10
          description: The number of items per page.
      responses:
        200:
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseSchemaBarangay'
    /api/v1/cityMunicipality:
    get:
      tags:
        - City Municipality
      summary: Get all cities and municipalities
      description: Returns a combined list of cities and municipalities.
      parameters:
        - in: query
          name: page
          schema:
            type: number
          description: Page number.
        - in: query
          name: limit
          schema:
            type: number
          description: Items per page.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: number
                  limit:
                    type: number
                  total:
                    type: number
                  data:
                    type: array
                    items:
                      oneOf:
                        - $ref: '#/components/schemas/CitySchema'
                        - $ref: '#/components/schemas/MunicipalitySchema'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseSchemaError'
...
  /api/v1/cityMunicipality/{code}/barangays:
    get:
      tags:
        - City/Municipality
      summary: Get all barangays in a city/municipality.
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
            example: 1380100000
          description: The 10-digit PSGC code of the city/municipality.
        - in: query
          name: page
          schema:
            type: number
            default: 1
          description: The page number of the response.
        - in: query
          name: limit
          schema:
            type: number
            default: 10
          description: The number of items per page.
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseSchemaBarangay'
...
components:
  schemas:
...
    ResponseSchemaBarangay:
      type: object
      properties:
        page:
          type: number
          description: The current page of the response.
        limit:
          type: number
          description: The number of items per page.
        total:
          type: number
          description: The total number of items.
        data:
          type: array
          description: The data of the response.
          items:
            $ref: '#/components/schemas/BarangaySchema'
    ResponseSchemaError:
      type: object
      properties:
        message:
          type: string
          description: The message of the response.
        error:
          type: boolean
          description: Whether the request is successful.
        status:
          type: number
          description: The HTTP status code of the response.
...
    CitySchema:
      type: object
      properties:
        psgc10DigitCode:
          type: string
          description: The 10-digit PSGC code of the city.
        name:
          type: string
          description: The name of the city.
        code:
          type: string
          description: The correspondence code of the city.
        geographicLevel:
          type: string
          description: The geographic level of the city.
        oldName:
          type: string
          description: The old name of the city.
        cityClass:
          type: string
          description: The class of the city.
        incomeClassification:
          type: string
          description: The income classification of the city.
        population2020:
          type: number
          description: The population count of the city in 2020.
        regionCode:
          type: string
          description: The 10-digit PSGC code of the region where the city belongs.
        provinceCode:
          type: string
          description: The 10-digit PSGC code of the province where the city belongs.
        isCapital:
          type: boolean
          description: Whether the city is a capital.
    MunicipalitySchema:
      type: object
      properties:
        psgc10DigitCode:
          type: string
          description: The 10-digit PSGC code of the municipality.
        name:
          type: string
          description: The name of the municipality.
        code:
          type: string
          description: The correspondence code of the municipality.
        geographicLevel:
          type: string
          description: The geographic level of the municipality.
        oldName:
          type: string
          description: The old name of the municipality.
        incomeClassification:
          type: string
          description: The income classification of the municipality.
        population2020:
          type: number
          description: The population count of the municipality in 2020.
        regionCode:
          type: string
          description: The 10-digit PSGC code of the region where the municipality belongs.
        provinceCode:
          type: string
          description: The 10-digit PSGC code of the province where the municipality belongs.
        isCapital:
          type: boolean
          description: Whether the municipality is a capital.
...
    BarangaySchema:
      type: object
      properties:
        psgc10DigitCode:
          type: string
          description: The 10-digit PSGC code of the barangay.
        name:
          type: string
          description: The name of the barangay.
        code:
          type: string
          description: The correspondence code of the barangay.
        geographicLevel:
          type: string
          description: The geographic level of the barangay.
        urbanRural:
          type: string
          description: The urban or rural status of the barangay.
        oldName:
          type: string
          description: The old name of the barangay.
        population2020:
          type: number
          description: The population count of the barangay in 2020.
        regionCode:
          type: string
          description: The 10-digit PSGC code of the region where the barangay belongs.
        provinceCode:
          type: string
          description: The 10-digit PSGC code of the province where the barangay belongs.
        cityMunicipalityCode:
          type: string
          description: The 10-digit PSGC code of the city or municipality where the barangay belongs.
        isPoblacion:
          type: boolean
          description: Whether the barangay is a poblacion.